Move helpers to test file
