--- thrifty-kotlin-codegen/src/test/kotlin/com/microsoft/thrifty/kgen/util.kt
@@ -1,15 +1,0 @@
-package com.microsoft.thrifty.kgen
-
-import com.squareup.kotlinpoet.FileSpec
-import io.kotest.matchers.compilation.shouldCompile
-
-fun List<FileSpec>.shouldCompile() {
-    forEach {
-        it.shouldCompile()
-    }
-}
-
-fun FileSpec.shouldCompile() {
-    toString().lineSequence().forEachIndexed { index, s -> println("${(index+1).toString().padStart(4)}: $s") }
-    toString().shouldCompile()
-}
